# api/docs/telephant-api-documentation.yml
openapi: 3.0.0
info:
  title: Telephant API
  description: |
    *The mechanism for swapping airtime is powered by Telephant's APIs. Developers can use the API to tap into the platform's capabilities and incorporate the service into their own apps.*

    The Telephant API's backbone is HTTP. The nature of our API is RESTful. It has a set of predictable resource URLs. It returns HTTP response codes to indicate errors. It also accepts JSON as an HTTP body payload and returns it, along with some other useful formats.
    
    You can use your favorite HTTP/REST library for your programming language or an HTTP client like Postman or CURL to use Telephant's API.

    Happy programming **:-)**

    ## HTTP Response Status Codes
    HTTP response status codes show whether or not a particular HTTP request was completed successfully. Section 10 of RFC 2616 defines the status codes listed below. RFC 7231 contains an updated specification.

    Telephant API makes use of the following HTTP status codes, grouped into three classes:

    ### Successful Responses - 2XX

    #### 200 - OK

    The request has succeeded.

    #### 201 - Created

    As a result of the request's success, a new resource has been generated. This is the response that is normally sent in response to POST or some PUT queries.

    ### Client Error - 4XX
    
    #### 400 - Bad Request
    Due to incorrect syntax, the server was unable to understand the request.
    
    #### 401 - Unauthorized
...
    
...
    
...
    
...

...

...

...

...

...
servers:
...
paths:
...
  /v1/airtime:
    get:
      summary: Get airtime balance.
      description: Check the balance of a group or individual accounts airtime balance.
      tags:
        - Airtime
      security:
        - BearerAuth: []
      parameters:
        - in: header
          name: account_id
          description: Clients can make requests as connected accounts using the special header account_id which should contain a Telephant account ID.
          schema:
            type: string
            format: uuid
          required: false
      responses:
        '200':
          description: Success
          content: 
            application/json:
              examples:
                Airtime-Get:
                  $ref: '#/components/examples/Airtime-Get'
        '4XX':
          $ref: '#/components/responses/4XX'

        '5XX':
          $ref: '#/components/responses/5XX'

    put:
      summary: Send airtime.
      description: |
        Send airtime credit to another Telephant user.

      tags:
        - Airtime
      security:
        - BearerAuth: []
      parameters:
        - in: header
          name: account_id
          description: Clients can make requests as connected accounts using the special header account_id which should contain a Telephant account ID.
          schema:
            type: string
            format: uuid
          required: false
      requestBody:
        description: |
          Define the details of the airtime send request.

          **Note:** Keep in mind that the following is required in order to send airtime successfully:
          * The balance in the sending accounting needs to exceed the amount to be sent.
          * You can send airtime to another Telephant account that is registered.
          * You can send airtime to an individual or a group account, but you must supply the correct account name.

        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - accountName
              properties:
                accountName:
                  type: string
                  description: Provide a valid Telephant account name.
                  examples:
                    - "27821234567"
                  required: true
                amount:
                  description: The amount of airtime to send to the recipient.
                  type: number
                  example: 10.00
      responses: 
        '200':
          description: Success
          content:
            application/json:
              example:
                { 
                  message: 'The airtime was sent successfully.' 
                }   
        '4XX':
          $ref: '#/components/responses/4XX'

        '5XX':
          $ref: '#/components/responses/5XX'

    post:
      summary: Use airtime.
      description: Convert positive airtime balance for use as airtime on phone.
      tags:
        - Airtime
      security:
        - BearerAuth: []
      responses:
        '4XX':
          $ref: '#/components/responses/4XX'

        '5XX':
          $ref: '#/components/responses/5XX'
  /v1/airtime/{ContactsPhoneNumber}/{Amount}:
    get:
      summary: Request airtime.
      description: |
        Request airtime credit from a user.

        **Note** This is a GET request so that POS labeling via QR codes can be supported. 
        A shopkeeper will be able to scan QR codes on the side of products and request credit 
        from the customer.
  
        This functionality could be added to the app. Where the shopkeeper first captures 
        the user's phone number, then scans each code and sends simple GET requests to 
        the server.  The functionality could be built into a laptop or tablet browser and would not 
        require sophisticated client side technology like scanners.
      tags:
        - Airtime
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: ContactsPhoneNumber
          schema:
            type: string
            pattern: '^(27)[0-9]{9}$'
          required: true
          description: A contact's unique mobile phone number.
        - in: path
          name: Amount
          schema:
            type: number
          required: true
          description: The amount of money you would like to request.
        - in: header
          name: account_id
          description: Clients can make requests as connected accounts using the special header account_id which should contain a Telephant account ID.
          schema:
            type: string
            format: uuid
          required: false
      responses:
        '4XX':
          $ref: '#/components/responses/4XX'

        '5XX':
          $ref: '#/components/responses/5XX'

#### CONTACTS ####
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    4XX:
      description: |
        Client error
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: Integer
                example: 400
                description: |
                  Please review the documentation for [HTTP Response Status Codes](/#section/HTTP-Response-Status-Codes)

                  Possible values are:
                    - 400 - Bad Request
                    - 401 - Unauthorized
                    - 403 - Forbidden
                    - 404 - Not Found
                    - 405 - Method Not Allowed
              name:
                type: String
                example: 'BadRequest'
                description: This value will always be returned as 'bumpy case'.
              message:
                type: String
                description: |
                  The message field will provide specific information as to the reason for the error.
                example: 'The account ID provided is not valid.'
    5XX:
      description: |
        Server error  
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: Integer
                example: 500
              name:
                type: String
                example: 'InternalServerError'
              message:
                type: String
                example: 'We are sorry but something went wrong on our side.'
  examples:
...
    Airtime-Get:
      value:
        {
          "accountID": "60ebd53e2f79311541618bc5",
          "accountName": "my-business-1",
          "description": "My Business Pty (Ltd)",
          "balance": 1448.85,
          "timestamp_ZA": "Sun Aug 01 2021 16:05:29 GMT+0200",
          "timestamp_UTC": "2021-08-01T14:05:29.042Z"
        }